requirements.txt
pandas==2.2.0
openpyxl==3.1.2
streamlit==1.38.0
streamlit-authenticator==0.4.2
Pillow==10.4.0
pyyaml

